pyproject.toml
[project]
name = "job-crawler"
version = "0.1.0"
description = "Discord bot that announces new job postings"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.4",
    "httpx>=0.28.1",
    "pydantic-ai[ollama]",
    "valkey>=6.0",
]

[project.scripts]
job-crawler = "job_crawler.bot:main"
job-preflight = "job_crawler.preflight:run"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]
